# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UP Dash - B2B fashion analytics intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients (admin only)
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated client list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedClients"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{clientId}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a single client by id
      parameters:
        - in: path
          name: clientId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: boolean
        code:
          type: string
        message:
          type: string
        status:
          type: integer
      required: [error, code, message, status]
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required: [email, password]
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required: [refreshToken]
    RefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
      required: [accessToken]
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [ADMIN, CLIENT]
        clientId:
          type: ["string", "null"]
      required: [id, email, firstName, lastName, role]
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"
      required: [accessToken, refreshToken, user]
    Client:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        apiKey:
          type: string
        revenueYtd:
          type: number
        ordersYtd:
          type: integer
        leadsYtd:
          type: integer
        approvedLeads:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - apiKey
        - revenueYtd
        - ordersYtd
        - leadsYtd
        - approvedLeads
        - isActive
        - createdAt
        - updatedAt
    PaginatedClients:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer
      required: [data, total, page, pages]
    CreateClientRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        apiKey:
          type: string
      required: [name, email, apiKey]
...