# lib/api-zod/src/generated/types/paginatedClients.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UP Dash - B2B fashion analytics intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { Client } from "./client";

export interface PaginatedClients {
  data: Client[];
  total: number;
  page: number;
  pages: number;
}